@@ -36,7 +36,7 @@ testlogger {
     showStandardStreams true
 }
 
-sourceCompatibility = '11'
+sourceCompatibility = '14'
 
 compileJava {
   options.release = 15
